Create separate EverythingServer projects for HTTP and stdio
