+++ samples/EverythingServer.Core/SubscriptionMessageSender.cs
@@ -1,8 +1,11 @@
 ﻿using System.Collections.Concurrent;
+using Microsoft.Extensions.Hosting;
 using ModelContextProtocol;
 using ModelContextProtocol.Server;
 
-internal class SubscriptionMessageSender(McpServer server, ConcurrentDictionary<string, byte> subscriptions) : BackgroundService
+namespace EverythingServer.Core;
+
+public class SubscriptionMessageSender(McpServer server, ConcurrentDictionary<string, byte> subscriptions) : BackgroundService
 {
     protected override async Task ExecuteAsync(CancellationToken stoppingToken)
     {
