+++ samples/EverythingServer.Core/Tools/AnnotatedMessageTool.cs
@@ -2,7 +2,7 @@
 using ModelContextProtocol.Server;
 using System.ComponentModel;
 
-namespace EverythingServer.Tools;
+namespace EverythingServer.Core.Tools;
 
 [McpServerToolType]
 public class AnnotatedMessageTool
