+++ samples/EverythingServer.Core/Tools/SampleLlmTool.cs
@@ -2,7 +2,7 @@
 using ModelContextProtocol.Server;
 using System.ComponentModel;
 
-namespace EverythingServer.Tools;
+namespace EverythingServer.Core.Tools;
 
 [McpServerToolType]
 public class SampleLlmTool
