+++ samples/EverythingServer.Http/Properties/launchSettings.json
